dus-backend/requirements.txt
Django==4.2.9
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
djongo==1.3.6
pymongo==3.12.3
dnspython==2.4.2
django-cors-headers==4.3.1
python-decouple==3.8
Pillow==10.2.0
anthropic>=0.40.0
celery==5.3.6
redis==5.0.1
django-filter==23.5
drf-spectacular==0.27.1
pymupdf>=1.23.0
django-storages>=1.14
boto3>=1.34.0
